Struct riddle_audio::ClipPlayerBuilder

pub struct ClipPlayerBuilder { /* fields omitted */ }

Builder for ClipPlayer

A builder instance may be used to construct multiple players.

Example

let bytes = include_bytes!("../../example_assets/boop.wav");
let clip = Clip::load(&bytes[..], ClipFormat::Wav)?;

// Play the clip
let player = ClipPlayerBuilder::new(&audio_system)
    .with_volume(0.5)
    .with_mode(PlayMode::Loop)
    .play(&clip)?;

Implementations

impl ClipPlayerBuilder

pub fn new(audio: &AudioSystem) -> Self

Make a new builder.

Defaults:

mode: PlayMode::OneShot

volume: 1.0.

pub fn with_mode(&mut self, mode: PlayMode) -> &mut Self

Set the playback mode of the player. Defaults to PlayMode::OneShot.

pub fn with_volume(&mut self, volume: f32) -> &mut Self

Set the playback volume of the player. Defaults to 1.0.

pub fn play(&self, clip: &Clip) -> Result<ClipPlayer, AudioError>

Build the ClipPlayer, and start playing the clip immediately.

pub fn paused(&self, clip: &Clip) -> Result<ClipPlayer, AudioError>

Build the ClipPlayer in the paused state. ClipPlayer::resume will need to be called on the player to start playback.